Fixed dimspeed slow test by replacing int with float Integrated a new Webpage updated several small fixes inserted tests for lock release inserted mutex for persistent data updated primary program to match functionality of test

+++ lib/NeoPixelManager/NeoPixelManager.h
@@ -26,7 +26,7 @@ private:
     bool smooth;
     int dataPin;
     int pixelcount;
-    int smoothStepwidth;
+    float smoothStepwidth;
     Color* ledTargetStateBuffer;
     Color* ledCurrentStateBuffer;
     void updateLEDStateFromBuffer();
